requirements.txt
pandas>=1.3.0
google-cloud-bigquery>=3.5.0
openpyxl>=3.0.10

gspread>=5.0.0
gspread_dataframe>=3.2.2
